feat: add AUR packaging, CI/release workflows, LICENSE, and code quality fixes
- Add MIT LICENSE file
- Update Cargo.toml metadata (correct repo URL, author)
- Update .gitignore to exclude packaging build artifacts
- Add .github/workflows/ci.yml (fmt, clippy, test, build)
- Add .github/workflows/release.yml (cross-compile x86_64/aarch64, GitHub Release, auto AUR push)
- Add aur/ramwise/PKGBUILD + .SRCINFO (build-from-source AUR package)
- Add aur/ramwise-bin/PKGBUILD + .SRCINFO (prebuilt binary AUR package)
- Add scripts/release.sh (bump version, tag, push)
- Add scripts/publish-aur.sh (clone AUR repo, update checksums, push)
- Update README.md with AUR install instructions
- Fix all clippy warnings: sort_by → sort_by_key, field_reassign_with_default
- Run cargo fmt across entire codebase

PKGBUILD verified locally: cargo build, cargo test (15/15 pass), package() all succeed

+++ src/history/buffer.rs
@@ -56,7 +56,7 @@ pub struct HistoryBuffer {
 
 impl HistoryBuffer {
     /// Create a new history buffer
-    /// 
+    ///
     /// # Arguments
     /// * `capacity` - Maximum number of snapshots to keep
     /// * `max_age` - Maximum age of snapshots before pruning
@@ -201,9 +201,7 @@ impl HistoryBuffer {
         let cutoff = now - duration;
 
         // Find the oldest point within the duration
-        let start_point = history
-            .iter()
-            .find(|p| p.timestamp >= cutoff)?;
+        let start_point = history.iter().find(|p| p.timestamp >= cutoff)?;
         let end_point = history.back()?;
 
         let actual_duration = end_point.timestamp.duration_since(start_point.timestamp);
@@ -292,10 +290,7 @@ impl HistoryBuffer {
 
     /// Get latest RSS for a process
     pub fn latest_rss(&self, pid: i32) -> Option<u64> {
-        self.process_history
-            .get(&pid)?
-            .back()
-            .map(|p| p.rss)
+        self.process_history.get(&pid)?.back().map(|p| p.rss)
     }
 }
 
@@ -345,7 +340,7 @@ mod tests {
                 processes: vec![ProcessMemory {
                     pid: 1234,
                     name: "leaky".to_string(),
-                    rss: 100_000_000 + (i as u64 * 10_000_000), // Growing by 10MB/sec
+                    rss: 100_000_000 + (i * 10_000_000), // Growing by 10MB/sec
                     ..Default::default()
                 }],
                 total_processes: 1,
